Be sure to save a file when releasing it
After mounting a cloudfuse file on a loopback device, the "losetup -d"
fails to send the block device writes back to cloudfiles.  It seems
the RELEASE operation only closes the local file buffer, without
uploading to to cloudfiles.

Change cfs_release to work like cfs_flush instead, which does not
close the local file buffer and does upload the changes to the file.
I think this is still suboptimal because if the image is only read
from, not written, I think this call will still be made, resulting
in an unnessary upload to cloudfiles.